Topology: Use uint64 to handle sizes instead of int
Related to https://trac.osgeo.org/postgis/ticket/4141

+++ liblwgeom/liblwgeom_topo.h
@@ -185,7 +185,6 @@ typedef struct LWT_BE_CALLBACKS_T {
    * @param ids an array of element identifiers
    * @param numelems input/output parameter, pass number of node identifiers
    *                 in the input array, gets number of node in output array.
-   *	TODO: Should be uint64 to match SPI_processed
    * @param fields fields to be filled in the returned structure, see
    *               LWT_COL_NODE_* macros
    *
@@ -198,7 +197,7 @@ typedef struct LWT_BE_CALLBACKS_T {
    */
   LWT_ISO_NODE* (*getNodeById) (
       const LWT_BE_TOPOLOGY* topo,
-      const LWT_ELEMID* ids, int* numelems, int fields
+      const LWT_ELEMID* ids, uint64_t* numelems, int fields
   );
 
   /**
@@ -210,7 +209,6 @@ typedef struct LWT_BE_CALLBACKS_T {
    * @param numelems output parameter, gets number of elements found
    *                 if the return is not null, otherwise see @return
    *                 section for semantic.
-   *	TODO: Should be uint64 to match SPI_processed
    * @param fields fields to be filled in the returned structure, see
    *               LWT_COL_NODE_* macros
    * @param limit max number of nodes to return, 0 for no limit, -1
@@ -224,7 +222,7 @@ typedef struct LWT_BE_CALLBACKS_T {
    */
   LWT_ISO_NODE* (*getNodeWithinDistance2D) (
       const LWT_BE_TOPOLOGY* topo,
-      const LWPOINT* pt, double dist, int* numelems,
+      const LWPOINT* pt, double dist, uint64_t* numelems,
       int fields, int limit
   );
 
@@ -238,14 +236,13 @@ typedef struct LWT_BE_CALLBACKS_T {
    * @param nodes the nodes to insert. Those with a node_id set to -1
    *              it will be replaced to an automatically assigned identifier
    * @param nelems number of elements in the nodes array
-   *	TODO: Should be uint64 to match SPI_processed
    *
    * @return 1 on success, 0 on error (@see lastErrorMessage)
    */
   int (*insertNodes) (
       const LWT_BE_TOPOLOGY* topo,
       LWT_ISO_NODE* nodes,
-      int numelems
+      uint64_t numelems
   );
 
   /**
@@ -266,7 +263,7 @@ typedef struct LWT_BE_CALLBACKS_T {
    */
   LWT_ISO_EDGE* (*getEdgeById) (
       const LWT_BE_TOPOLOGY* topo,
-      const LWT_ELEMID* ids, int* numelems, int fields
+      const LWT_ELEMID* ids, uint64_t* numelems, int fields
   );
 
   /**
@@ -291,7 +288,7 @@ typedef struct LWT_BE_CALLBACKS_T {
    */
   LWT_ISO_EDGE* (*getEdgeWithinDistance2D) (
       const LWT_BE_TOPOLOGY* topo,
-      const LWPOINT* pt, double dist, int* numelems,
+      const LWPOINT* pt, double dist, uint64_t* numelems,
       int fields, int limit
   );
 
@@ -319,14 +316,13 @@ typedef struct LWT_BE_CALLBACKS_T {
    * @param edges the edges to insert. Those with a edge_id set to -1
    *              it will be replaced to an automatically assigned identifier
    * @param nelems number of elements in the edges array
-   *	TODO: Should be uint64 to match SPI_processed
    *
    * @return number of inserted edges, or -1 (@see lastErrorMessage)
    */
   int (*insertEdges) (
       const LWT_BE_TOPOLOGY* topo,
       LWT_ISO_EDGE* edges,
-      int numelems
+      uint64_t numelems
   );
 
   /**
@@ -372,7 +368,7 @@ typedef struct LWT_BE_CALLBACKS_T {
    */
   LWT_ISO_FACE* (*getFaceById) (
       const LWT_BE_TOPOLOGY* topo,
-      const LWT_ELEMID* ids, int* numelems, int fields
+      const LWT_ELEMID* ids, uint64_t* numelems, int fields
   );
 
   /**
@@ -451,7 +447,7 @@ typedef struct LWT_BE_CALLBACKS_T {
   LWT_ISO_NODE* (*getNodeWithinBox2D) (
       const LWT_BE_TOPOLOGY* topo,
       const GBOX* box,
-      int* numelems, int fields, int limit
+      uint64_t* numelems, int fields, int limit
   );
 
   /**
@@ -476,7 +472,7 @@ typedef struct LWT_BE_CALLBACKS_T {
   LWT_ISO_EDGE* (*getEdgeWithinBox2D) (
       const LWT_BE_TOPOLOGY* topo,
       const GBOX* box,
-      int* numelems, int fields, int limit
+      uint64_t* numelems, int fields, int limit
   );
 
   /**
@@ -499,7 +495,7 @@ typedef struct LWT_BE_CALLBACKS_T {
    */
   LWT_ISO_EDGE* (*getEdgeByNode) (
       const LWT_BE_TOPOLOGY* topo,
-      const LWT_ELEMID* ids, int* numelems, int fields
+      const LWT_ELEMID* ids, uint64_t* numelems, int fields
   );
 
   /**
@@ -556,14 +552,14 @@ typedef struct LWT_BE_CALLBACKS_T {
    * @param faces the faces to insert. Those with a node_id set to -1
    *              it will be replaced to an automatically assigned identifier
    * @param nelems number of elements in the faces array
-   *	TODO: Should be uint64 to match SPI_processed
+   *	TODO: Should be uint64_t to match SPI_processed
    *
    * @return number of inserted faces, or -1 (@see lastErrorMessage)
    */
   int (*insertFaces) (
       const LWT_BE_TOPOLOGY* topo,
       LWT_ISO_FACE* faces,
-      int numelems
+      uint64_t numelems
   );
 
   /**
@@ -573,7 +569,7 @@ typedef struct LWT_BE_CALLBACKS_T {
    * @param faces an array of LWT_ISO_FACE object with selecting id
    *              and setting mbr.
    * @param numfaces number of faces in the "faces" array
-   *	TODO: Should be uint64 to match SPI_processed
+   *	TODO: Should be uint64_t to match SPI_processed
    *
    * @return number of faces being updated or -1 on error
    *         (@see lastErroMessage)
@@ -612,7 +608,7 @@ typedef struct LWT_BE_CALLBACKS_T {
    */
   LWT_ELEMID* (*getRingEdges) (
       const LWT_BE_TOPOLOGY* topo,
-      LWT_ELEMID edge, int *numedges, int limit
+      LWT_ELEMID edge, uint64_t *numedges, uint64_t limit
   );
 
   /**
@@ -622,7 +618,7 @@ typedef struct LWT_BE_CALLBACKS_T {
    * @param edges an array of LWT_ISO_EDGE object with selecting id
    *              and updating fields.
    * @param numedges number of edges in the "edges" array
-   *	TODO: Should be uint64 to match SPI_processed
+   *	TODO: Should be uint64_t to match SPI_processed
    * @param upd_fields fields to be updated for the selected edges,
    *                   see LWT_COL_EDGE_* macros
    *
@@ -631,7 +627,7 @@ typedef struct LWT_BE_CALLBACKS_T {
    */
   int (*updateEdgesById) (
       const LWT_BE_TOPOLOGY* topo,
-      const LWT_ISO_EDGE* edges, int numedges,
+      const LWT_ISO_EDGE* edges, uint64_t numedges,
       int upd_fields
   );
 
@@ -646,7 +642,6 @@ typedef struct LWT_BE_CALLBACKS_T {
    *                 in the input array, gets number of edges in output array
    *                 if the return is not null, otherwise see @return
    *                 section for semantic.
-   *	TODO: Should be uint64 to match SPI_processed
    * @param fields fields to be filled in the returned structure, see
    *               LWT_COL_EDGE_* macros
    * @param box optional bounding box to further restrict matches, use
@@ -658,7 +653,7 @@ typedef struct LWT_BE_CALLBACKS_T {
    */
   LWT_ISO_EDGE* (*getEdgeByFace) (
       const LWT_BE_TOPOLOGY* topo,
-      const LWT_ELEMID* ids, int* numelems, int fields,
+      const LWT_ELEMID* ids, uint64_t* numelems, int fields,
       const GBOX *box
   );
 
@@ -671,7 +666,7 @@ typedef struct LWT_BE_CALLBACKS_T {
    *                 identifiers in the input array, gets number of
    *                 nodes in output array if the return is not null,
    *                 otherwise see @return section for semantic.
-   *	TODO: Should be uint64 to match SPI_processed
+   *	TODO: Should be uint64_t to match SPI_processed
    * @param fields fields to be filled in the returned structure, see
    *               LWT_COL_NODE_* macros
    * @param box optional bounding box to further restrict matches, use
@@ -683,7 +678,7 @@ typedef struct LWT_BE_CALLBACKS_T {
    */
   LWT_ISO_NODE* (*getNodeByFace) (
       const LWT_BE_TOPOLOGY* topo,
-      const LWT_ELEMID* faces, int* numelems, int fields,
+      const LWT_ELEMID* faces, uint64_t* numelems, int fields,
       const GBOX *box
   );
 
@@ -694,7 +689,6 @@ typedef struct LWT_BE_CALLBACKS_T {
    * @param nodes an array of LWT_ISO_EDGE objects with selecting id
    *              and updating fields.
    * @param numnodes number of nodes in the "nodes" array
-   *	TODO: Should be uint64 to match SPI_processed
    * @param upd_fields fields to be updated for the selected edges,
    *                   see LWT_COL_NODE_* macros
    *
@@ -703,7 +697,7 @@ typedef struct LWT_BE_CALLBACKS_T {
    */
   int (*updateNodesById) (
       const LWT_BE_TOPOLOGY* topo,
-      const LWT_ISO_NODE* nodes, int numnodes,
+      const LWT_ISO_NODE* nodes, uint64_t numnodes,
       int upd_fields
   );
 
@@ -713,7 +707,6 @@ typedef struct LWT_BE_CALLBACKS_T {
    * @param topo the topology to act upon
    * @param ids an array of face identifiers
    * @param numelems number of face identifiers in the ids array
-   *	TODO: Should be uint64 to match SPI_processed
    *
    * @return number of faces being deleted or -1 on error
    *         (@see lastErrorMessage)
@@ -721,7 +714,7 @@ typedef struct LWT_BE_CALLBACKS_T {
   int (*deleteFacesById) (
       const LWT_BE_TOPOLOGY* topo,
       const LWT_ELEMID* ids,
-      int numelems
+      uint64_t numelems
   );
 
   /**
@@ -753,7 +746,6 @@ typedef struct LWT_BE_CALLBACKS_T {
    * @param topo the topology to act upon
    * @param ids an array of node identifiers
    * @param numelems number of node identifiers in the ids array
-   *	TODO: Should be uint64 to match SPI_processed
    *
    * @return number of nodes being deleted or -1 on error
    *         (@see lastErrorMessage)
@@ -761,7 +753,7 @@ typedef struct LWT_BE_CALLBACKS_T {
   int (*deleteNodesById) (
       const LWT_BE_TOPOLOGY* topo,
       const LWT_ELEMID* ids,
-      int numelems
+      uint64_t numelems
   );
 
   /**
@@ -857,7 +849,6 @@ typedef struct LWT_BE_CALLBACKS_T {
    * @param numelems output parameter, gets number of elements found
    *                 if the return is not null, otherwise see @return
    *                 section for semantic.
-   *	TODO: Should be uint64 to match SPI_processed
    * @param fields fields to be filled in the returned structure, see
    *               LWT_COL_FACE_* macros
    * @param limit max number of faces to return, 0 for no limit, -1
@@ -872,7 +863,7 @@ typedef struct LWT_BE_CALLBACKS_T {
   LWT_ISO_FACE* (*getFaceWithinBox2D) (
       const LWT_BE_TOPOLOGY* topo,
       const GBOX* box,
-      int* numelems, int fields, int limit
+      uint64_t* numelems, int fields, uint64_t limit
   );
 
 } LWT_BE_CALLBACKS;
